auto-tag: add conservative performer aliases auto tagging
This initial implementation is based just on a rough guess on data I've seen so far of matches against
FreeOnes names/aliases for performers where there are many one word aliases and sometimes names.
FreeOnes-Babe "Belinda" is one example of a performer that is more widely known
by here aliases that are also more "robust", in contrast the name "Belinda" which is not judged to be robust
by this current implementation, would produce a lot false-positives.

The autotag.tagger has now an Aliases field in addition to the Name.
Both Aliases and Name are no longer strings but QualifiedNames which have the meta information Qualified
to indicate if the name or aliases are judged good for "robust" matches.

For the creation of QualifiedNames some helper function have been added in a new
autotag/tag_utils file.

This initial change to auto-tagging will only use qualified/"robust" names
but later implementations should add "possible" matches with for example getPossiblePerformerMatches
for performers_possible_scenes, performers_possible_images ...
These possible matches should than be presented separately in the UI with the possibility to move a
"possible" match to a normal match.
The "possible" matching logic needs to make sure the matches are unique over "possible" and normal "robust" matches.